+++ FirebaseVertexAI/Sample/ChatSample/ViewModels/ConversationViewModel.swift
@@ -21,8 +21,8 @@ class ConversationViewModel: ObservableObject {
   /// This array holds both the user's and the system's chat messages
   @Published var messages = [ChatMessage]()
 
-  /// Indicates we're waiting for the model to finish
-  @Published var busy = false
+  /// Indicates we're waiting for the model to finish or the UI is loading
+  @Published var busy = true
 
   @Published var error: Error?
   var hasError: Bool {
@@ -30,18 +30,20 @@ class ConversationViewModel: ObservableObject {
   }
 
   private var model: GenerativeModel
-  private var chat: Chat
+  private var chat: Chat? = nil
   private var stopGenerating = false
 
   private var chatTask: Task<Void, Never>?
 
   init() {
     model = VertexAI.vertexAI().generativeModel(modelName: "gemini-1.5-flash")
-    chat = model.startChat()
+    Task {
+      await startNewChat()
+    }
   }
 
   func sendMessage(_ text: String, streaming: Bool = true) async {
-    error = nil
+    stop()
     if streaming {
       await internalSendMessageStreaming(text)
     } else {
@@ -49,11 +51,14 @@ class ConversationViewModel: ObservableObject {
     }
   }
 
-  func startNewChat() {
+  func startNewChat() async {
+    busy = true
+    defer {
+      busy = false
+    }
     stop()
-    error = nil
-    chat = model.startChat()
     messages.removeAll()
+    chat = await model.startChat()
   }
 
   func stop() {
@@ -62,8 +67,6 @@ class ConversationViewModel: ObservableObject {
   }
 
   private func internalSendMessageStreaming(_ text: String) async {
-    chatTask?.cancel()
-
     chatTask = Task {
       busy = true
       defer {
@@ -79,7 +82,10 @@ class ConversationViewModel: ObservableObject {
       messages.append(systemMessage)
 
       do {
-        let responseStream = chat.sendMessageStream(text)
+        guard let chat else {
+          throw ChatError.notInitialized
+        }
+        let responseStream = try await chat.sendMessageStream(text)
         for try await chunk in responseStream {
           messages[messages.count - 1].pending = false
           if let text = chunk.text {
@@ -95,8 +101,6 @@ class ConversationViewModel: ObservableObject {
   }
 
   private func internalSendMessage(_ text: String) async {
-    chatTask?.cancel()
-
     chatTask = Task {
       busy = true
       defer {
@@ -112,10 +116,12 @@ class ConversationViewModel: ObservableObject {
       messages.append(systemMessage)
 
       do {
-        var response: GenerateContentResponse?
-        response = try await chat.sendMessage(text)
+        guard let chat = chat else {
+          throw ChatError.notInitialized
+        }
+        let response = try await chat.sendMessage(text)
 
-        if let responseText = response?.text {
+        if let responseText = response.text {
           // replace pending message with backend response
           messages[messages.count - 1].message = responseText
           messages[messages.count - 1].pending = false
@@ -126,5 +132,9 @@ class ConversationViewModel: ObservableObject {
         messages.removeLast()
       }
     }
+  }
+
+  enum ChatError: Error {
+    case notInitialized
   }
 }